% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcvTT.R
\docType{data}
\name{nlcvTT}
\alias{nlcvTT}
\title{Data to Demonstrate nlcv and Co Functions}
\format{
The object is of class \code{"nlcv"}, an object as produced
by the \code{nlcv} function.
}
\source{
data simulated using:
 \code{nlcvTT <- nlcv(selBcrAblOrNeg, classVar = 'mol.biol', 
 classdist = "unbalanced", nRuns = 10, fsMethod = "t.test", 
verbose = TRUE)}
}
\usage{
nlcvTT
}
\description{
Simulated data set used to demonstrate nlcv and accompanying
plot functions to study classification problems
}
\examples{
\dontrun{
data(nlcvTT)
if (require(nlcv)) # on R-Forge
 scoresPlot(nlcvTT, tech = 'svm', nfeat = 25)
}
}
\seealso{
\code{\link[nlcv]{nlcv}}
}
\keyword{datasets}
