% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{oaPalette}
\alias{oaPalette}
\title{Generate a Palette of OA Colors}
\usage{
oaPalette(numColors = NULL, alpha = 1)
}
\arguments{
\item{numColors}{number of colors to be contained in the palette}

\item{alpha}{transparency level of the colors}
}
\value{
vector of colors
}
\description{
Generate a Palette of OA Colors
}
\author{
Jason Waddell
}
