% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSmoothScat.R
\name{plotCombMultSamples}
\alias{plotCombMultSamples}
\title{Plots the correlation in gene expression between more than 2 samples}
\usage{
plotCombMultSamples(exprsMatrix, ...)
}
\arguments{
\item{exprsMatrix}{ExpressionSet object to plot. For larger datasets,
this will typically be a subset of the data.}

\item{...}{Further arguments, e.g. to add extra plot options. See \code{\link{pairs}}}
}
\value{
no returned value, a plots is drawn in the current device
}
\description{
Plots the correlation in gene expression between more than 2 samples
}
\examples{
if (require(ALL)){
 data(ALL, package = "ALL")
 ALL <- addGeneInfo(ALL)
 plotCombMultSamples(exprs(ALL)[,c("84004", "11002", "01003")])
}
}
\seealso{
\code{\link{plotComb2Samples}}
}
\author{
Willem Talloen
}
