% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propDEgenes.R
\name{propDEgenes}
\alias{propDEgenes}
\alias{propDEgenes,limma-method}
\alias{propDEgenes,numeric-method}
\title{Generic function to compute the proportion of differentially expressed
  genes that are present}
\usage{
propDEgenes(object, ...)

\S4method{propDEgenes}{limma}(object, ...)

\S4method{propDEgenes}{numeric}(object, ...)
}
\arguments{
\item{object}{object of class \code{propDEgene}}

\item{...}{further arguments for the method (currently none implemented)}
}
\value{
numeric of length one giving the proportion of differentially expressed genes
}
\description{
Generic function to compute the proportion of differentially expressed
  genes that are present
}
\section{Methods}{

\describe{
limma
\item{object = "limma"}{propDEgenes method for a limma object}
numeric
\item{object = "numeric"}{propDEgenes method for a numeric vector, i.e. a vector
 of P Values}
}
}

\seealso{
\code{\link{propDEgenes-methods}}
}
\author{
Willem Talloen and Tobias Verbeke
}
\keyword{htest}
