% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propDEgenes.R
\name{propDEgenes}
\alias{propDEgenes}
\title{Generic function to compute the proportion of differentially expressed
  genes that are present}
\usage{
propDEgenes(object, ...)
}
\arguments{
\item{object}{object of class \code{propDEgene}}

\item{...}{further arguments for the method (currently none implemented)}
}
\value{
numeric of length one giving the proportion of differentially expressed genes
}
\description{
Generic function to compute the proportion of differentially expressed
  genes that are present
}
\author{
Willem Talloen and Tobias Verbeke
}
\keyword{htest}
