% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicates}
\alias{replicates}
\title{computes replicates across a vector}
\usage{
replicates(x)
}
\arguments{
\item{x}{character or numeric vector}
}
\value{
numeric vector
}
\description{
Given a vector, returns the replicates in order
}
\examples{
  
x <- c('a','b','a','a','b','a','c','c','c')
  data.frame(val=x,rep=replicates(x))
}
\references{
R-help mailing list
}
\seealso{
\code{\link{rle}}
}
\author{
Henrique Dallazuanna
}
\keyword{manip}
