% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R, R/tTest.R
\docType{methods}
\name{topTable,limma-method}
\alias{topTable,limma-method}
\alias{topTable,MArrayLM-method}
\alias{topTable-methods}
\alias{topTable,tTest-method}
\alias{topTable,fTest-method}
\title{Methods for topTable}
\usage{
\S4method{topTable}{limma}(
  fit,
  n = 10,
  coef = 2,
  genelist = fit$genes,
  eb = fit[c("t", "p.value", "lods")],
  adjust.method = "BH",
  sort.by = "B",
  resort.by = NULL,
  p.value = 1,
  lfc = 0
)

\S4method{topTable}{MArrayLM}(
  fit,
  n,
  coef = 2,
  genelist = fit$genes,
  eb = fit[c("t", "p.value", "lods")],
  adjust.method = "BH",
  sort.by = "B",
  resort.by = NULL,
  p.value = 1,
  lfc = 0
)

\S4method{topTable}{tTest}(fit, n)

\S4method{topTable}{fTest}(fit, n)
}
\arguments{
\item{fit}{object resulting from a classification or regression procedure}

\item{n}{number of features that one wants to extract from a table that
ranks all features according to their importance in the classification
or regression model; defaults to 10 for limma objects}
}
\description{
Methods for topTable. topTable extracts the top n most important features
for a given classification or regression procedure 
\section{Methods}{
\describe{
glmnet
\item{fit = "glmnet", n = "numeric"}{glmnet objects are produced by \code{lassoClass} or \code{lassoReg}}
limma
\item{fit = "limma", n = "numeric"}{limma objects are produced by \code{limma2Groups}}
MarrayLM
\item{fit = "limma", n = "numeric"}{MarrayLM objects are produced by \code{lmFit} of the \code{limma package}}
pamClass
\item{fit = "pamClass", n = "numeric"}{pamClass objects are produced by \code{pamClass}}
rfClass
\item{fit = "rfClass", n = "numeric"}{rfClass objects are produced by \code{rfClass}}
tTest
\item{fit = "tTest", n = "numeric"}{tTest objects are produced by \code{tTest}}
fTest
\item{fit = "fTest", n = "numeric"}{fTest objects are produced by \code{fTest}}
}
}
}
\seealso{
\code{\link[a4Core]{topTable,glmnet-method}},
\code{\link[a4Core]{topTable,lognet-method}},
\code{\link[a4Core]{topTable,elnet-method}},
\code{\link[a4Classif]{pamClass,lognet-method}},
\code{\link[a4Classif]{rfClass,elnet-method}},
}
\keyword{manip}
\keyword{methods}
