% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{limmaReg}
\alias{limmaReg}
\title{Wrapper for the limma function for the comparison of two groups 
(two factor levels)}
\usage{
limmaReg(object, covariable, probe2gene = TRUE)
}
\arguments{
\item{object}{object of class ExpressionSet}

\item{covariable}{string indicating the variable defining the continuous covariate}

\item{probe2gene}{logical; if \code{TRUE} Affymetrix probeset IDs are translated
into gene symbols; if \code{FALSE} no such translation is done}
}
\description{
Wrapper for the limma function for the comparison of two groups 
(two factor levels)
}
